large/pyt/requirements.txt
telegram_send==0.34
feedparser==6.0.10
schedule==1.1.0
beautifulsoup4==4.11.1
html-telegraph-poster==0.5.0
requests
pykakasi
tqdm
wiktionaryparser
nltk
pygtrans
python-telegram-bot==13.11